removed listener from map fragment map fragment extends PermissionFragment.kt
