+++ app/src/main/java/com/rootstrap/android/ui/activity/main/targetpoint/TargetPointsActivity.kt
@@ -5,11 +5,9 @@ import android.os.Bundle
 import com.rootstrap.android.R
 import com.rootstrap.android.databinding.ActivityTargetPointsBinding
 import com.rootstrap.android.ui.activity.main.authentication.ProfileActivity
-import com.rootstrap.android.util.permissions.PermissionActivity
-import com.rootstrap.android.util.permissions.PermissionResponse
-import com.rootstrap.android.util.permissions.locationPermissions
+import com.rootstrap.android.ui.base.BaseActivity
 
-class TargetPointsActivity : PermissionActivity(), MapFragment.MapFragmentInteraction {
+class TargetPointsActivity : BaseActivity() {
 
     private lateinit var binding: ActivityTargetPointsBinding
 
@@ -35,19 +33,5 @@ class TargetPointsActivity : PermissionActivity(), MapFragment.MapFragmentIntera
     private fun initMapFragment() {
         val fragment = MapFragment.newInstance()
         supportFragmentManager.beginTransaction().replace(R.id.fragment_map, fragment).commit()
-    }
-
-    override fun askForLocationPermission(permissionGranted: () -> Unit) {
-        requestPermission(
-            locationPermissions,
-            object : PermissionResponse {
-                override fun granted() {
-                    permissionGranted()
-                }
-
-                override fun denied() = Unit
-
-                override fun foreverDenied() = Unit
-            })
     }
 }
